pyproject.toml
[project]
name = "215-project-part-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "formdt>=0.4.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "psycopg2>=2.9.11",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.8",
    "sqlfluff>=3.5.0",
]

[tool.ruff]
line-length = 88

[tool.setuptools]
packages = ["pipeline"]

[tool.pytest]
minversion = "9.0"
addopts = ["-ra", "-q", "--cov", "--cov-report=term-missing"]
testpaths = [
    "tests",
]
pythonpath = [
    ".",
]
python_files = ["test_*.py", "*_test.py"]

[tool.coverage.run]
source = [
    "pipeline"
]
omit = [
    "*/__init__.py"
]

[tool.coverage.report]
show_missing = true
